# server/api/src/openapi_server/openapi/openapi.yaml
openapi: 3.0.1
info:
  description: Kaldi Customization Server.
  title: Kaldi Customization Server
  version: 1.1.0
servers:
- url: http://localhost:8080/api/v1
- url: https://localhost:8080/api/v1
...
paths:
  /audio:
    get:
      operationId: get_all_audio
      responses:
        200:
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Audio'
                type: array
          description: List of resources
        403:
          content: {}
          description: Forbidden
      security:
      - oauth:
        - read:audio
      summary: Returns a list of available audio
      tags:
      - decode
      x-openapi-router-controller: openapi_server.controllers.decode_controller
    post:
      operationId: upload_audio
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/BinaryResourceObject'
        required: true
      responses:
        201:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Audio'
          description: Created Resource
        403:
          content: {}
          description: Forbidden
        405:
          content: {}
          description: Invalid Input
      security:
      - oauth:
        - write:audio
      summary: Uploads audio
      tags:
      - decode
      x-openapi-router-controller: openapi_server.controllers.decode_controller
  /audio/{audio_uuid}:
    delete:
      description: Delete a single audio resource
      operationId: delete_audio_by_uuid
      parameters:
      - description: UUID of audio to delete
        explode: false
        in: path
        name: audio_uuid
        required: true
        schema:
          type: string
        style: simple
      responses:
        200:
          content: {}
          description: Success
        403:
          content: {}
          description: Forbidden
        404:
          content: {}
          description: Audio not found
      security:
      - oauth:
        - write:audio
      summary: Delete audio by UUID
      tags:
      - decode
      x-openapi-router-controller: openapi_server.controllers.decode_controller
    get:
      description: Returns a single audio resource
      operationId: get_audio_by_uuid
      parameters:
      - description: UUID of audio to return
        explode: false
        in: path
        name: audio_uuid
        required: true
        schema:
          type: string
        style: simple
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Audio'
          description: Audio object
        403:
          content: {}
          description: Forbidden
        404:
          content: {}
          description: Audio not found
      security:
      - oauth:
        - read:audio
      summary: Find audio by UUID
      tags:
      - decode
      x-openapi-router-controller: openapi_server.controllers.decode_controller
  /audio/{audio_uuid}/data:
    get:
      description: Returns the audio content
      operationId: get_audio_data
      parameters:
      - description: UUID of resource to return
        explode: false
        in: path
        name: audio_uuid
        required: true
        schema:
          type: string
        style: simple
      responses:
        200:
          content:
            audio/wav:
              schema:
                format: binary
                type: string
          description: Uploaded audio content
        403:
          content: {}
          description: Forbidden
        404:
          content: {}
          description: Audio not found
      security:
      - oauth:
        - read:audio
      summary: Returns the audio content
      tags:
      - decode
      x-openapi-router-controller: openapi_server.controllers.decode_controller
...
  /project/{project_uuid}/training/{training_version}/decode:
    get:
      description: Returns a list of all decodings for this training version
      operationId: get_decodings
      parameters:
      - description: UUID of the project
        explode: false
        in: path
        name: project_uuid
        required: true
        schema:
          example: 550e8400-e29b-11d4-a716-446655440000
          format: uuid
          type: string
        style: simple
      - description: Training version of the project
        explode: false
        in: path
        name: training_version
        required: true
        schema:
          type: integer
        style: simple
      responses:
        200:
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/DecodeAudio'
                type: array
          description: List of decoding messages
        400:
          content: {}
          description: Training not finished
        403:
          content: {}
          description: Forbidden
        404:
          content: {}
          description: Project or training not found
      security:
      - oauth:
        - decode:projects
      summary: List of all decodings
      tags:
      - decode
      x-openapi-router-controller: openapi_server.controllers.decode_controller
    post:
      description: Assign audio to current decoding session
      operationId: assign_audio_to_current_session
      parameters:
      - description: UUID of the project
        explode: false
        in: path
        name: project_uuid
        required: true
        schema:
          example: 550e8400-e29b-11d4-a716-446655440000
          format: uuid
          type: string
        style: simple
      - description: Training version of the project
        explode: false
        in: path
        name: training_version
        required: true
        schema:
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AudioReferenceObject'
        description: Audio that needs to be decoded
        required: true
      responses:
        201:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DecodeAudio'
          description: Audio successfully added to session
        403:
          content: {}
          description: Forbidden
        404:
          content: {}
          description: Project or training not found or no active session
      security:
      - oauth:
        - decode:projects
      summary: Assign Audio to decoding session
      tags:
      - decode
      x-openapi-router-controller: openapi_server.controllers.decode_controller
  /project/{project_uuid}/training/{training_version}/decode/session:
    delete:
      description: Delete the active decoding session
      operationId: delete_decode_session
      parameters:
      - description: UUID of the project
        explode: false
        in: path
        name: project_uuid
        required: true
        schema:
          example: 550e8400-e29b-11d4-a716-446655440000
          format: uuid
          type: string
        style: simple
      - description: Training version of the project
        explode: false
        in: path
        name: training_version
        required: true
        schema:
          type: integer
        style: simple
      responses:
        200:
          content: {}
          description: Session deleted
        403:
          content: {}
          description: Forbidden
        404:
          content: {}
          description: Project or training not found or no active decoding session
      security:
      - oauth:
        - decode:projects
      summary: Delete the decoding session
      tags:
      - decode
      x-openapi-router-controller: openapi_server.controllers.decode_controller
    get:
      description: Get the current decode session
      operationId: get_all_decode_sessions
      parameters:
      - description: UUID of the project
        explode: false
        in: path
        name: project_uuid
        required: true
        schema:
          example: 550e8400-e29b-11d4-a716-446655440000
          format: uuid
          type: string
        style: simple
      - description: Training version of the project
        explode: false
        in: path
        name: training_version
        required: true
        schema:
          type: integer
        style: simple
      responses:
        200:
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Resource'
                type: array
          description: The decode session for this training
        403:
          content: {}
          description: Forbidden
        404:
          content: {}
          description: Project or training not found
      security:
      - oauth:
        - decode:projects
      summary: Get the all sessions
      tags:
      - decode
      x-openapi-router-controller: openapi_server.controllers.decode_controller
    post:
      description: Create a new decoding session
      operationId: create_decode_session
      parameters:
      - description: UUID of the project
        explode: false
        in: path
        name: project_uuid
        required: true
        schema:
          example: 550e8400-e29b-11d4-a716-446655440000
          format: uuid
          type: string
        style: simple
      - description: Training version of the project
        explode: false
        in: path
        name: training_version
        required: true
        schema:
          type: integer
        style: simple
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DecodeSession'
          description: The new active DecodeSession
        400:
          content: {}
          description: An active session already exists
        403:
          content: {}
          description: Forbidden
        404:
          content: {}
          description: Project or training not found
      security:
      - oauth:
        - decode:projects
      summary: Create a new decoding session
      tags:
      - decode
      x-openapi-router-controller: openapi_server.controllers.decode_controller
  /project/{project_uuid}/training/{training_version}/decode/session/current:
    get:
      description: Get the current decode session
      operationId: get_current_decode_session
      parameters:
      - description: UUID of the project
        explode: false
        in: path
        name: project_uuid
        required: true
        schema:
          example: 550e8400-e29b-11d4-a716-446655440000
          format: uuid
          type: string
        style: simple
      - description: Training version of the project
        explode: false
        in: path
        name: training_version
        required: true
        schema:
          type: integer
        style: simple
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DecodeSession'
          description: The active DecodeSession
        403:
          content: {}
          description: Forbidden
        404:
          content: {}
          description: Project or training not found or session not started
      security:
      - oauth:
        - decode:projects
      summary: Get the current session
      tags:
      - decode
      x-openapi-router-controller: openapi_server.controllers.decode_controller
  /project/{project_uuid}/training/{training_version}/decode/session/{session_uuid}:
    get:
      description: Gets a specified session
      operationId: get_decode_session
      parameters:
      - description: UUID of the project
        explode: false
        in: path
        name: project_uuid
        required: true
        schema:
          example: 550e8400-e29b-11d4-a716-446655440000
          format: uuid
          type: string
        style: simple
      - description: Training version of the project
        explode: false
        in: path
        name: training_version
        required: true
        schema:
          type: integer
        style: simple
      - description: UUID of the session
        explode: false
        in: path
        name: session_uuid
        required: true
        schema:
          example: 550e8400-e29b-11d4-a716-446655440000
          format: uuid
          type: string
        style: simple
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DecodeSession'
          description: The queried decode session
        403:
          content: {}
          description: Forbidden
        404:
          content: {}
          description: Project, training or session not found
      security:
      - oauth:
        - decode:projects
      summary: Get a decode session
      tags:
      - decode
      x-openapi-router-controller: openapi_server.controllers.decode_controller
  /project/{project_uuid}/training/{training_version}/decode/session/{session_uuid}/commit:
    put:
      description: Enqueue the currently active session for decoding
      operationId: start_decode
      parameters:
      - description: UUID of the project
        explode: false
        in: path
        name: project_uuid
        required: true
        schema:
          example: 550e8400-e29b-11d4-a716-446655440000
          format: uuid
          type: string
        style: simple
      - description: Training version of the project
        explode: false
        in: path
        name: training_version
        required: true
        schema:
          type: integer
        style: simple
      - description: UUID of the session
        explode: false
        in: path
        name: session_uuid
        required: true
        schema:
          example: 550e8400-e29b-11d4-a716-446655440000
          format: uuid
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CallbackObject'
        description: Callbackobject that gets executed after process
        required: false
      responses:
        202:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DecodeSession'
          description: Session successfully commited
        400:
          content: {}
          description: Training not finished or session already in progress
        403:
          content: {}
          description: Forbidden
        404:
          content: {}
          description: Project or training not found
      security:
      - oauth:
        - decode:projects
      summary: Commits the decode session for decoding
      tags:
      - decode
      x-openapi-router-controller: openapi_server.controllers.decode_controller
...
  /project/{project_uuid}/training/{training_version}/decode/{audio_uuid}:
    delete:
      description: Unassign audio to current decoding session
      operationId: unassign_audio_to_current_session
      parameters:
      - description: UUID of the project
        explode: false
        in: path
        name: project_uuid
        required: true
        schema:
          example: 550e8400-e29b-11d4-a716-446655440000
          format: uuid
          type: string
        style: simple
      - description: Training version of the project
        explode: false
        in: path
        name: training_version
        required: true
        schema:
          type: integer
        style: simple
      - description: UUID of the audio
        explode: false
        in: path
        name: audio_uuid
        required: true
        schema:
          example: 550e8400-e29b-11d4-a716-446655440000
          format: uuid
          type: string
        style: simple
      responses:
        201:
          content: {}
          description: Audio successfully unassigned from session
        403:
          content: {}
          description: Forbidden
        404:
          content: {}
          description: Project or training not found or no active session or audio
            not in active session
      security:
      - oauth:
        - decode:projects
      summary: Unassign Audio to decoding session
      tags:
      - decode
      x-openapi-router-controller: openapi_server.controllers.decode_controller
    get:
      description: Returns the result of a decoding task
      operationId: get_decode_result
      parameters:
      - description: UUID of the project
        explode: false
        in: path
        name: project_uuid
        required: true
        schema:
          example: 550e8400-e29b-11d4-a716-446655440000
          format: uuid
          type: string
        style: simple
      - description: Training version of the project
        explode: false
        in: path
        name: training_version
        required: true
        schema:
          type: integer
        style: simple
      - description: UUID of the audio
        explode: false
        in: path
        name: audio_uuid
        required: true
        schema:
          example: 550e8400-e29b-11d4-a716-446655440000
          format: uuid
          type: string
        style: simple
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DecodeAudio'
          description: Result of the decoding task
        204:
          content: {}
          description: Decoding not yet completed
        403:
          content: {}
          description: Forbidden
        404:
          content: {}
          description: Project or training not found
        415:
          content: {}
          description: Unsupported media type
      security:
      - oauth:
        - decode:projects
      summary: Get the result of a decoding task
      tags:
      - decode
      x-openapi-router-controller: openapi_server.controllers.decode_controller
...
components:
  schemas:
...
    DecodeAudio:
      example:
        transcripts:
        - '{}'
        - '{}'
        session_uuid: 550e8400-e29b-11d4-a716-446655440000
      properties:
        transcripts:
          items:
            properties:
              text:
                type: string
              confidence:
                type: number
          type: array
        audio:
          $ref: '#/components/schemas/Audio'
        session_uuid:
          example: 550e8400-e29b-11d4-a716-446655440000
          format: uuid
          type: string
      type: object
...
    Resource:
      example:
        creation_timestamp: 2000-01-23T04:56:07.000+00:00
        name: myFile.pdf
        uuid: 550e8400-e29b-11d4-a716-446655440000
      properties:
        uuid:
          example: 550e8400-e29b-11d4-a716-446655440000
          format: uuid
          type: string
        name:
          example: myFile.pdf
          type: string
        status:
          $ref: '#/components/schemas/ResourceStatus'
        resource_type:
          $ref: '#/components/schemas/ResourceType'
        creation_timestamp:
          format: date-time
          type: string
      required:
      - name
      - resource_type
      - status
      - uuid
      type: object
    Audio:
      example:
        creation_timestamp: 2000-01-23T04:56:07.000+00:00
        name: myAudio.wav
        uuid: 550e8400-e29b-11d4-a716-446655440000
      properties:
        uuid:
          example: 550e8400-e29b-11d4-a716-446655440000
          format: uuid
          type: string
        name:
          example: myAudio.wav
          type: string
        creation_timestamp:
          format: date-time
          type: string
        status:
          $ref: '#/components/schemas/AudioStatus'
      required:
      - name
      - status
      - uuid
      type: object
...
    ResourceType:
      enum:
      - 1
      - 2
      - 3
      - 4
      - 5
      - 6
      type: integer
      x-enum-varnames:
      - html
      - docx
      - txt
      - pdf
      - png
      - jpg
...
    AudioStatus:
      enum:
      - 100
      - 150
      - 200
      - 300
      - 320
      type: integer
      x-enum-varnames:
      - Init
      - AudioPrep_Pending
      - AudioPrep_In_Progress
      - AudioPrep_Success
      - AudioPrep_Failure
...
    AudioReferenceObject:
      example:
        audio_uuid: 550e8400-e29b-11d4-a716-446655440000
      properties:
        audio_uuid:
          description: UUID of the audio file
          example: 550e8400-e29b-11d4-a716-446655440000
          format: uuid
          type: string
      required:
      - audio_uuid
      type: object
...
    BinaryResourceObject:
      properties:
        upfile:
          description: File object that needs to be uploaded
          format: binary
          type: string
      required:
      - upfile
      type: object
    CallbackObject:
      example:
        url: url
      properties:
        url:
          description: The URL to be called
          type: string
        method:
          $ref: '#/components/schemas/CallbackMethod'
      required:
      - method
      - url
  securitySchemes:
    oauth:
      flows:
        password:
          scopes:
            write:user: modify the user account
            read:user: read the user account
            write:projects: modify projects in your account
            read:projects: read your projects
            train:projects: train projects in your account
            decode:projects: decode audio data using projects in your account
            write:resources: modify resources in your account
            read:resources: read your resources
            write:audio: modify your audio
            read:audio: read your audio
          tokenUrl: http://localhost:8080/api/oauth/token
      type: oauth2
      x-tokenInfoFunc: openapi_server.controllers.security_controller_.info_from_oauth
      x-scopeValidateFunc: openapi_server.controllers.security_controller_.validate_scope_oauth